+++ src/industry-solution-packs-composable-applications/contracts/solution-operations-governance-assurance-profile.js
@@ -1,0 +1,50 @@
+export class SolutionOperationsGovernanceAssuranceProfile {
+  constructor({
+    profileName,
+    ownershipFields = [],
+    serviceProfileFields = [],
+    serviceObjectiveFields = [],
+    runbookFields = [],
+    incidentFields = [],
+    rolloutFields = [],
+    evidenceFields = [],
+    lifecycleStates = [],
+    controls = [],
+    failureRecovery = [],
+    observabilityFields = [],
+    assuranceActivities = [],
+    invariants = [],
+    ownershipAssigned = true,
+    serviceProfilesVersioned = true,
+    objectivesDefined = true,
+    observabilityPayloadSafe = true,
+    runbooksTested = true,
+    incidentsEvidencePreserving = true,
+    rolloutTenantScoped = true,
+    evidenceImmutable = true,
+    assuranceIndependent = true,
+    providerAuthorityPreserved = true,
+    tenantIsolated = true,
+    sharedOwnership = false,
+    serviceProfileUnversioned = false,
+    telemetryCopiesProtectedPayload = false,
+    runbookUntested = false,
+    incidentFabricatesState = false,
+    bigBangTenantRollout = false,
+    rolloutWithoutRollback = false,
+    mutableEvidence = false,
+    selfAssurance = false,
+    extensionWeakensControls = false,
+    clientInfersCompletion = false,
+    aiSelfPromotes = false
+  } = {}) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({ ownershipFields, serviceProfileFields, serviceObjectiveFields, runbookFields, incidentFields, rolloutFields, evidenceFields, lifecycleStates, controls, failureRecovery, observabilityFields, assuranceActivities, invariants })) {
+      this[key] = Object.freeze([...value]);
+    }
+    for (const [key, value] of Object.entries({ ownershipAssigned, serviceProfilesVersioned, objectivesDefined, observabilityPayloadSafe, runbooksTested, incidentsEvidencePreserving, rolloutTenantScoped, evidenceImmutable, assuranceIndependent, providerAuthorityPreserved, tenantIsolated, sharedOwnership, serviceProfileUnversioned, telemetryCopiesProtectedPayload, runbookUntested, incidentFabricatesState, bigBangTenantRollout, rolloutWithoutRollback, mutableEvidence, selfAssurance, extensionWeakensControls, clientInfersCompletion, aiSelfPromotes })) {
+      this[key] = Boolean(value);
+    }
+    Object.freeze(this);
+  }
+}